Refactor: symmetric Moved aes_sym and shake_sym modules into symmetric
